backend/requirements.txt
flask
flask-cors
torch
torchvision
transformers
numpy
pillow
opencv-python-headless
easyocr
sentencepiece
pycld2